requirements.txt
Django==4.0.6
mysqlclient==2.1.1
djangorestframework==3.13.1
drf-yasg==1.20.0